Add files via upload
In this tutorial I will try to describe in more detail how to use the program, as well as save the topic on the promotion of HF and HF queries with this program.

First, run the program as an administrator, activate your license. Then we minimize the program and open the accounts.txt file. There, we specify the accounts in the column login:password format (separated by a colon between the login and password). Let's close the file and save it, keeping in mind the number of accounts.  Then we open the urls folder and create in it as many TXT files, as many accounts. I.e. for each account TXT file with posts, posts/links should be in columns (new line - new post). For example, if you want to post the same links to 100 accounts, then create in the folder urls TXT file with links and duplicate it 100 times (a simple example, see video: http://www.youtube.com/watch?v=MU-eiMAC330). Well, if there is a ТХТ file with 300 links, you can split this TXT file into 3 parts and create 3 ТхТ documents in each urls folder with 100 links. Then again, go into the program. We press the button URLS and specify the PATH to the folder urls. Next, click on accounts and specify the TXT file with accounts. Just in case, specify the Antigate key. The delay is recommended to put in a range of 200-300 seconds, so that there are no problems with limits. If you do not post links, and just text or links with text, then MUST put a tick "Do not parse TITLE", otherwise there will be errors! If you decide to specify a proxy, then specify only working and fast proxy! Now you can press the "Posting" button.

Now let's talk about promoting BF and HF queries with this program. In fact, everything is very simple, but it takes a long time and it is desirable to have a server for this. In addition to this promoted page should be well optimized in terms of SEO. 
And also for this need for 50-100 akcahs a week with a quickerobotom, it is best to use the downloaded akcah, because they are cheap (where to buy them see below). The application is simple: take the page and drive it through the acres throughout the week, then change the acres. And so until we get the right positions. Operates such a thing only for Yasha, works for a couple of years.